requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cppimport==22.8.2
cuda-toolkit==13.1.0
cupy-cuda12x==13.6.0
cycler==0.12.1
dllist==2.0.0
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.12
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cuda-runtime==13.1.80
nvidia-cuda-runtime-cu12==12.9.79
nvidia-ml-py==13.590.44
nvidia-modelopt==0.40.0
onnx==1.19.1
onnx-ir==0.1.13
onnx_graphsurgeon==0.5.8
onnxconverter-common==1.16.0
onnxruntime-directml==1.20.0
onnxscript==0.5.7
onnxslim==0.1.82
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
polygraphy==0.49.26
protobuf==6.33.2
psutil==7.2.1
PuLP==3.3.0
pybind11==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorrt==10.13.0.35
tensorrt_cu12==10.13.0.35
tensorrt_cu12_bindings==10.13.0.35
tensorrt_cu12_libs==10.13.0.35
tensorrt_cu13==10.13.3.9.post1
tensorrt_cu13_bindings==10.13.3.9.post1
tensorrt_cu13_libs==10.13.3.9.post1
tokenizers==0.22.1
torch==2.9.1+cu128
torch_tensorrt==2.9.0+cu128
torchprofile==0.0.4
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.3
triton-windows==3.5.1.post23
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.7
ultralytics-thop==2.0.18
urllib3==2.6.2
wxPython==4.2.4
